+++ frontend/app/src/components/settings/OnboardingSettings.vue
@@ -8,19 +8,19 @@ import {
   required
 } from '@vuelidate/validators';
 import { type Ref } from 'vue';
+import isEqual from 'lodash/isEqual';
 import { type BackendOptions } from '@/electron-main/ipc';
 import { type Writeable } from '@/types';
 import { LogLevel } from '@/utils/log-level';
-import {
-  type BackendConfiguration,
-  type DefaultBackendArguments
-} from '@/types/backend';
+import { type BackendConfiguration } from '@/types/backend';
+
+const { t } = useI18n();
 
 const emit = defineEmits<{
   (e: 'dismiss'): void;
 }>();
 
-const { dataDirectory } = storeToRefs(useMainStore());
+const { dataDirectory, defaultBackendArguments } = storeToRefs(useMainStore());
 
 const userDataDirectory: Ref<string> = ref('');
 const userLogDirectory: Ref<string> = ref('');
@@ -28,7 +28,7 @@ const logFromOtherModules: Ref<boolean> = ref(false);
 const maxLogSize: Ref<number> = ref(0);
 const sqliteInstructions: Ref<number> = ref(0);
 const maxLogFiles: Ref<number> = ref(0);
-const formValid: Ref<boolean> = ref(false);
+const valid: Ref<boolean> = ref(false);
 
 const { backendSettings } = useSettingsApi();
 
@@ -38,45 +38,40 @@ const configuration: Ref<BackendConfiguration> = asyncComputed(() =>
   backendSettings()
 );
 
-const { info } = useInfoApi();
-const defaultArguments: Ref<DefaultBackendArguments> = asyncComputed(
-  () => info().then(info => info.backendDefaultArguments),
-  {
-    maxLogfilesNum: 0,
-    maxSizeInMbAllLogs: 0,
-    sqliteInstructions: 0
-  }
-);
-
-const { t } = useI18n();
-
-const loaded = async () => {
+const initialOptions: ComputedRef<Partial<BackendOptions>> = computed(() => {
   const config = get(configuration);
   const opts = get(options);
-  const defaults = get(defaultArguments);
-
-  set(userDataDirectory, opts.dataDirectory ?? get(dataDirectory));
-  set(userLogDirectory, opts.logDirectory ?? get(defaultLogDirectory));
-  set(logLevel, opts.loglevel ?? get(defaultLogLevel));
-  set(logFromOtherModules, opts.logFromOtherModules ?? false);
-  set(
-    maxLogFiles,
-    opts.maxLogfilesNum ??
-      config.maxLogfilesNum.value ??
-      defaults.maxLogfilesNum
-  );
-  set(
-    maxLogSize,
-    opts.maxSizeInMbAllLogs ??
-      config.maxSizeInMbAllLogs.value ??
-      defaults.maxSizeInMbAllLogs
-  );
-  set(
-    sqliteInstructions,
-    opts.sqliteInstructions ??
-      config.sqliteInstructions.value ??
+  const defaults = get(defaultBackendArguments);
+  return {
+    loglevel: opts.loglevel ?? get(defaultLogLevel),
+    dataDirectory: opts.dataDirectory ?? get(dataDirectory),
+    logDirectory: opts.logDirectory ?? get(defaultLogDirectory),
+    logFromOtherModules: opts.logFromOtherModules ?? false,
+    maxLogfilesNum:
+      opts.maxLogfilesNum ??
+      config?.maxLogfilesNum?.value ??
+      defaults.maxLogfilesNum,
+    maxSizeInMbAllLogs:
+      opts.maxSizeInMbAllLogs ??
+      config?.maxSizeInMbAllLogs?.value ??
+      defaults.maxSizeInMbAllLogs,
+    sqliteInstructions:
+      opts.sqliteInstructions ??
+      config?.sqliteInstructions?.value ??
       defaults.sqliteInstructions
-  );
+  };
+});
+
+const loaded = async () => {
+  const initial = get(initialOptions);
+
+  set(logLevel, initial.loglevel);
+  set(userDataDirectory, initial.dataDirectory);
+  set(userLogDirectory, initial.logDirectory);
+  set(logFromOtherModules, initial.logFromOtherModules);
+  set(maxLogFiles, initial.maxLogfilesNum);
+  set(maxLogSize, initial.maxSizeInMbAllLogs);
+  set(sqliteInstructions, initial.sqliteInstructions);
 };
 
 const {
@@ -90,22 +85,22 @@ const {
 } = useBackendManagement(loaded);
 
 const isMaxLogFilesDefault = computed(() => {
-  const defaults = get(defaultArguments);
+  const defaults = get(defaultBackendArguments);
   return defaults.maxLogfilesNum === get(maxLogFiles);
 });
 
 const isMaxSizeDefault = computed(() => {
-  const defaults = get(defaultArguments);
+  const defaults = get(defaultBackendArguments);
   return defaults.maxSizeInMbAllLogs === get(maxLogSize);
 });
 
 const isSqliteInstructionsDefaults = computed(() => {
-  const defaults = get(defaultArguments);
+  const defaults = get(defaultBackendArguments);
   return defaults.sqliteInstructions === get(sqliteInstructions);
 });
 
 const resetDefaultArguments = (field: 'files' | 'size' | 'instructions') => {
-  const defaults = get(defaultArguments);
+  const defaults = get(defaultBackendArguments);
   if (field === 'files') {
     set(maxLogFiles, defaults.maxLogfilesNum);
   } else if (field === 'size') {
@@ -116,60 +111,55 @@ const resetDefaultArguments = (field: 'files' | 'size' | 'instructions') => {
 };
 
 const newUserOptions = computed(() => {
-  const options: Writeable<Partial<BackendOptions>> = {};
+  const initial = get(initialOptions);
+  const newOptions: Writeable<Partial<BackendOptions>> = {};
+
   const level = get(logLevel);
-  const defaultLevel = get(defaultLogLevel);
-  if (level !== defaultLevel) {
-    options.loglevel = level;
-  }
-
-  const userLog = get(userLogDirectory);
-  if (userLog !== get(defaultLogDirectory)) {
-    options.logDirectory = userLog;
-  }
-
-  const logFromOther = get(logFromOtherModules);
-  if (logFromOther) {
-    options.logFromOtherModules = true;
+  if (level !== initial.loglevel) {
+    newOptions.loglevel = level;
   }
 
   const userData = get(userDataDirectory);
-  if (userData !== get(dataDirectory)) {
-    options.dataDirectory = userData;
+  if (userData !== initial.dataDirectory) {
+    newOptions.dataDirectory = userData;
   }
 
-  if (isDefined(defaultArguments)) {
-    const defaults = get(defaultArguments);
-    if (defaults.maxLogfilesNum !== get(maxLogFiles)) {
-      options.maxLogfilesNum = get(maxLogFiles);
-    } else {
-      delete options.maxLogfilesNum;
-    }
-
-    if (defaults.maxSizeInMbAllLogs !== get(maxLogSize)) {
-      options.maxSizeInMbAllLogs = get(maxLogSize);
-    } else {
-      delete options.maxSizeInMbAllLogs;
-    }
-
-    if (defaults.sqliteInstructions !== get(sqliteInstructions)) {
-      options.sqliteInstructions = get(sqliteInstructions);
-    } else {
-      delete options.sqliteInstructions;
-    }
+  const userLog = get(userLogDirectory);
+  if (userLog !== initial.logDirectory) {
+    newOptions.logDirectory = userLog;
   }
 
-  return options;
+  const logFromOther = get(logFromOtherModules);
+  if (logFromOther !== initial.logFromOtherModules) {
+    newOptions.logFromOtherModules = logFromOther;
+  }
+  if (!get(isMaxLogFilesDefault)) {
+    newOptions.maxLogfilesNum = get(maxLogFiles);
+  }
+
+  if (!get(isMaxSizeDefault)) {
+    newOptions.maxSizeInMbAllLogs = get(maxLogSize);
+  }
+
+  if (!get(isSqliteInstructionsDefaults)) {
+    newOptions.sqliteInstructions = get(sqliteInstructions);
+  }
+
+  return newOptions;
 });
 
-const valid = computed(() => {
-  const newOptions = get(newUserOptions);
-  const oldOptions = get(options);
-  const defaultLevel = get(defaultLogLevel);
-  const newLogLevel = newOptions.loglevel ?? defaultLevel;
-  const oldLogLevel = oldOptions.loglevel ?? defaultLevel;
-  const updatedKeys = Object.keys(newOptions).filter(s => s !== 'loglevel');
-  return updatedKeys.length > 0 || newLogLevel !== oldLogLevel;
+const anyValueChanged = computed(() => {
+  const form: Partial<BackendOptions> = {
+    loglevel: get(logLevel),
+    dataDirectory: get(userDataDirectory),
+    logDirectory: get(userLogDirectory),
+    logFromOtherModules: get(logFromOtherModules),
+    maxSizeInMbAllLogs: get(maxLogSize),
+    sqliteInstructions: get(sqliteInstructions),
+    maxLogfilesNum: get(maxLogFiles)
+  };
+
+  return !isEqual(form, get(initialOptions));
 });
 
 const { openDirectory } = useInterop();
@@ -202,7 +192,7 @@ const v$ = useVuelidate(
 );
 
 watch(v$, ({ $invalid }) => {
-  set(formValid, !$invalid);
+  set(valid, !$invalid);
 });
 
 const icon = (level: LogLevel): string => {
@@ -287,20 +277,6 @@ const showResetConfirmation = () => {
     reset
   );
 };
-
-watch(defaultArguments, args => {
-  if (!get(maxLogSize) && args.maxSizeInMbAllLogs) {
-    set(maxLogSize, args.maxSizeInMbAllLogs);
-  }
-
-  if (!get(maxLogFiles) && args.maxLogfilesNum) {
-    set(maxLogFiles, args.maxLogfilesNum);
-  }
-
-  if (!get(sqliteInstructions) && args.sqliteInstructions) {
-    set(sqliteInstructions, args.sqliteInstructions);
-  }
-});
 </script>
 
 <template>
@@ -326,6 +302,7 @@ watch(defaultArguments, args => {
 
     <v-text-field
       v-model="userDataDirectory"
+      data-cy="user-data-directory-input"
       :loading="!userDataDirectory"
       class="pt-2"
       outlined
@@ -351,9 +328,10 @@ watch(defaultArguments, args => {
       </template>
     </v-text-field>
 
-    <v-form :value="formValid">
+    <v-form :value="valid">
       <v-text-field
         v-model="userLogDirectory"
+        data-cy="user-log-directory-input"
         :disabled="!!fileConfig.logDirectory"
         :persistent-hint="!!fileConfig.logDirectory"
         :hint="
@@ -376,6 +354,7 @@ watch(defaultArguments, args => {
       <v-select
         v-model="logLevel"
         :items="levels"
+        class="loglevel-input"
         :disabled="!!fileConfig.loglevel"
         :label="t('backend_settings.settings.log_level.label')"
         :persistent-hint="!!fileConfig.loglevel"
@@ -406,12 +385,15 @@ watch(defaultArguments, args => {
 
       <v-expansion-panels flat>
         <v-expansion-panel>
-          <v-expansion-panel-header>
+          <v-expansion-panel-header
+            data-cy="onboarding-setting__advance-toggle"
+          >
             {{ t('backend_settings.advanced') }}
           </v-expansion-panel-header>
           <v-expansion-panel-content>
             <v-text-field
               v-model.number="maxLogSize"
+              data-cy="max-log-size-input"
               outlined
               :hint="
                 !!fileConfig.maxSizeInMbAllLogs
@@ -421,19 +403,21 @@ watch(defaultArguments, args => {
               :label="t('backend_settings.max_log_size.label')"
               :disabled="fileConfig.maxSizeInMbAllLogs"
               :persistent-hint="!!fileConfig.maxSizeInMbAllLogs"
-              :loading="!configuration || !defaultArguments"
+              :loading="!configuration || !defaultBackendArguments"
               :error-messages="v$.maxLogSize.$errors.map(e => e.$message)"
               type="number"
             >
               <template #append>
                 <setting-reset-button
                   v-if="!isMaxSizeDefault"
+                  data-cy="reset-max-log-size"
                   @click="resetDefaultArguments('size')"
                 />
               </template>
             </v-text-field>
             <v-text-field
               v-model.number="maxLogFiles"
+              data-cy="max-log-files-input"
               outlined
               :hint="t('backend_settings.max_log_files.hint')"
               :label="
@@ -443,13 +427,14 @@ watch(defaultArguments, args => {
               "
               :disabled="fileConfig.maxLogfilesNum"
               :persistent-hint="!!fileConfig.maxLogfilesNum"
-              :loading="!configuration || !defaultArguments"
+              :loading="!configuration || !defaultBackendArguments"
               :error-messages="v$.maxLogFiles.$errors.map(e => e.$message)"
               type="number"
             >
               <template #append>
                 <setting-reset-button
                   v-if="!isMaxLogFilesDefault"
+                  data-cy="reset-max-log-files"
                   @click="resetDefaultArguments('files')"
                 />
               </template>
@@ -457,6 +442,7 @@ watch(defaultArguments, args => {
 
             <v-text-field
               v-model.number="sqliteInstructions"
+              data-cy="sqlite-instructions-input"
               outlined
               :hint="
                 !!fileConfig.sqliteInstructions
@@ -466,7 +452,7 @@ watch(defaultArguments, args => {
               :label="t('backend_settings.sqlite_instructions.label')"
               :disabled="fileConfig.sqliteInstructions"
               :persistent-hint="!!fileConfig.sqliteInstructions"
-              :loading="!configuration || !defaultArguments"
+              :loading="!configuration || !defaultBackendArguments"
               :error-messages="
                 v$.sqliteInstructions.$errors.map(e => e.$message)
               "
@@ -475,6 +461,7 @@ watch(defaultArguments, args => {
               <template #append>
                 <setting-reset-button
                   v-if="!isSqliteInstructionsDefaults"
+                  data-cy="reset-sqlite-instructions"
                   @click="resetDefaultArguments('instructions')"
                 />
               </template>
@@ -482,6 +469,7 @@ watch(defaultArguments, args => {
 
             <v-checkbox
               v-model="logFromOtherModules"
+              data-cy="log-from-other-modules-checkbox"
               :label="t('backend_settings.log_from_other_modules.label')"
               :disabled="fileConfig.logFromOtherModules"
               persistent-hint
@@ -506,8 +494,9 @@ watch(defaultArguments, args => {
       </v-btn>
       <v-btn
         depressed
+        data-cy="onboarding-setting__submit-button"
         color="primary"
-        :disabled="!valid || !formValid"
+        :disabled="!anyValueChanged || !valid"
         @click="save()"
       >
         {{ t('common.actions.save') }}