+++ frontend/app/src/composables/backend.ts
@@ -52,7 +52,6 @@ export const useBackendManagement = (loaded: () => void = () => {}) => {
   const restartBackendWithOptions = async (
     options: Partial<BackendOptions>
   ) => {
-    const { setConnected, connect } = useMainStore();
     await setConnected(false);
     await interop.restartBackend(options);
     await connect();
@@ -78,15 +77,17 @@ export const useBackendManagement = (loaded: () => void = () => {}) => {
       loglevel,
       ...opts
     };
-    saveUserOptions(updatedOptions);
-    set(userOptions, updatedOptions);
+    await applyUserOptions(updatedOptions);
+  };
+
+  const applyUserOptions = async (config: Partial<BackendOptions>) => {
+    saveUserOptions(config);
+    set(userOptions, config);
     await restartBackendWithOptions(get(options));
   };
 
   const resetOptions = async () => {
-    saveUserOptions({});
-    set(userOptions, {});
-    await restartBackendWithOptions(get(options));
+    await applyUserOptions({});
   };
 
   const restartBackend = async () => {